Adds test for RifFileProgressTracker and moves it to more appropriate package.
